requirements/local.txt
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-model-utils==4.2.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
Unipath==1.1